requirements.txt
numpy==2.1.2
Pillow==10.4.0
Requests==2.32.3
sentence_transformers==3.2.0
streamlit==1.39.0
transformers==4.45.2
torch==2.4.1+cpu